libs/shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sql2ai-shared"
version = "0.1.0"
description = "SQL2.AI Shared Enterprise Library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Chris Therriault", email = "chris@servicevision.net" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "aioodbc>=0.5.0",

    # Redis
    "redis>=5.0.0",

    # HTTP
    "httpx>=0.26.0",

    # Resilience
    "tenacity>=8.2.0",
    "circuitbreaker>=2.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Observability
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
    "opentelemetry-instrumentation-redis>=0.43b0",
    "structlog>=24.1.0",

    # Utilities
    "ulid-py>=1.1.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
]

ai = [
    "litellm>=1.20.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["sql2ai_shared"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
